+++ src/crypt.encoders/Crc32Encoder.cs
@@ -2,16 +2,15 @@
 /// Implementation of the `Crypt.Encoders.Crc32Encoder` class.
 
 namespace Crypt.Encoders {
+  using System;
   using System.Globalization;
-  
-  using MiniFramework.Security.Cryptography;
-  using Properties;
+  using System.Text;
+
+  using Crypt.Encoders.Internal;
+  using Crypt.Encoders.Properties;
 
   /// Represents the CRC32 encoding method.
   public class Crc32Encoder: IStringEncoder {
-  
-    /// Initializes a new instance of the class.
-    public Crc32Encoder() {}
 
     /// @property Description
     /// The encoder description.
@@ -29,7 +28,9 @@ namespace Crypt.Encoders {
     /// @param text The string to encode.
     /// @returns The encoded string.
     public string Encode(string text) {
-      return HashUtility.ComputeCrc32(text).ToString(CultureInfo.InvariantCulture);
+      var buffer = Encoding.Default.GetBytes(text);
+      var hash = Crc32.Create().ComputeHash(buffer);
+      return BitConverter.ToUInt32(hash, 0).ToString(CultureInfo.InvariantCulture);
     }
   }
 }