+++ src/crypt.encoders/TripleDesEncoder.cs
@@ -6,13 +6,10 @@ namespace Crypt.Encoders {
   using System.Security.Cryptography;
   using System.Text;
 
-  using Properties;
+  using Crypt.Encoders.Properties;
   
   /// Represents the TripleDES encoding method.
   public class TripleDesEncoder: IStringEncoder {
-  
-    /// Initializes a new instance of the class.
-    public TripleDesEncoder() {}
 
     /// @property Description
     /// The encoder description.
@@ -30,8 +27,9 @@ namespace Crypt.Encoders {
     /// @param text The string to encode.
     /// @returns The encoded string.
     public string Encode(string text) {
-      var bytes = Encoding.Default.GetBytes(text);
-      return Convert.ToBase64String(TripleDES.Create().CreateEncryptor().TransformFinalBlock(bytes, 0, bytes.Length));
+      var buffer = Encoding.Default.GetBytes(text);
+      var hash = TripleDES.Create().CreateEncryptor().TransformFinalBlock(buffer, 0, buffer.Length);
+      return Convert.ToBase64String(hash);
     }
   }
 }