+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This file contains highlights of what changes on each version of the [Crypt.cs](
 
 #### Version 0.5.2
 - Added support for [SonarQube](http://www.sonarqube.org) code analyzer.
+- Removed dependency on `MiniFramework` library.
 
 #### Version 0.5.1
 - Replaced the custom build scripts by [NAnt](http://nant.sourceforge.net).